reso-metadata-utils: move the metadata-report serializer in (Stage 2a of #221)
Move the EDMX → metadata-report serializer (serializeMetadataReport,
generateMetadataReport + the MetadataReport* types) from reso-certification
into reso-metadata-utils's main (browser-safe) index. Pure transform, no
new deps.

- Drop the serializer re-exports from cert's metadata barrel + public index
  (no pass-through); repoint every consumer — cert metadata/sdk/cli + mcp —
  directly at reso-metadata-utils.
- Move the serializer test in (86 tests in reso-metadata-utils now).
- Drop reso-metadata-utils's unused reso-common dep: nothing in the package
  imports it (csdl, serializer, and the coming XSD/fetchers don't) — it's a
  sibling, not a substrate. Corrected the index header + package description.

The XSD validator (2b/2c) stays isolated in cert for now — revisit after
more deps move, which may dissolve the original cross-dep build issue.

Verified: full root precommit green (lint + typecheck + every package).
