+++ reso-certification/src/sdk/dd.ts
@@ -11,8 +11,8 @@ import { existsSync } from 'node:fs';
 import { join, dirname } from 'node:path';
 import { resolveAuthToken } from '../test-runner/auth.js';
 import { fetchMetadataWithVersion, persistMetadataXml } from '../test-runner/metadata.js';
-import { generateMetadataReport } from '../metadata/serializer.js';
-import type { MetadataReport } from '../metadata/serializer.js';
+import { generateMetadataReport } from '@reso-standards/reso-metadata-utils';
+import type { MetadataReport } from '@reso-standards/reso-metadata-utils';
 import { runDdMetadataChecks } from '../metadata/dd-metadata-checks.js';
 import type { DdReference } from '../metadata/dd-metadata-checks.js';
 import { fetchAndMergeLookupResource } from '../metadata/lookup-resource.js';
