+++ reso-mcp-server/package.json
@@ -23,6 +23,7 @@
     "@modelcontextprotocol/sdk": "^1.29.0",
     "zod": "^3.23.0",
     "@reso-standards/reso-client": "file:../reso-client",
+    "@reso-standards/reso-metadata-utils": "file:../reso-metadata-utils",
     "@reso-standards/reso-validation": "file:../reso-validation",
     "@reso-standards/reso-certification": "file:../reso-certification",
     "@reso-standards/odata-expression-parser": "file:../odata-expression-parser"
